modified:   app/Http/Controllers/BarangController.php 	modified:   app/Http/Controllers/LoginController.php 	modified:   app/Http/Controllers/RegisterController.php 	modified:   app/Models/User.php 	modified:   public/css/login.css 	modified:   public/css/registration.css 	new file:   resources/views/cart.blade.php 	modified:   resources/views/index.blade.php 	modified:   resources/views/login.blade.php 	modified:   resources/views/product.blade.php 	new file:   resources/views/products.blade.php 	modified:   resources/views/register.blade.php 	modified:   routes/web.php

+++ app/Http/Controllers/BarangController.php
@@ -106,17 +106,23 @@ class BarangController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(string $id)
     {
         $barang = Barang::find($id);
-        //delete image
-        Storage::delete('public/foto/'. $barang->foto);
 
-        //delete post
+        // Check if barang exists
+        if (!$barang) {
+            return redirect()->route('product')->with(['error' => 'Data not found']);
+        }
+
+        // Delete image
+        Storage::delete('public/foto/' . $barang->foto);
+
+        // Delete post
         $barang->delete();
 
-        //redirect to index
-        return redirect()->route('product')
-        ->with(['success' => 'Data Berhasil Dihapus!']);
+        // Redirect to index
+        return redirect()->route('product')->with(['success' => 'Data Berhasil Dihapus!']);
     }
+
 }
